+++ nemo/collections/common/parts/__init__.py
@@ -12,5 +12,5 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from nemo.collections.common_parts.parts.multi_layer_perveptron import MultiLayerPerceptron
-from nemo.collections.common_parts.parts.transformer_utils import *
+from nemo.collections.common.parts.multi_layer_perceptron import MultiLayerPerceptron
+from nemo.collections.common.parts.transformer_utils import *
