+++ nemo/collections/nlp/models/ner_model.py
@@ -23,14 +23,14 @@ from torch.utils.data import DataLoader
 # TODO replace with nemo module
 from transformers import BertModel
 
-from nemo.collections.common_parts.parts.multi_layer_perveptron import MultiLayerPerceptron
-from nemo.collections.common_parts.parts.transformer_utils import transformer_weights_init
-from nemo.collections.common_parts.tokenizers.bert_tokenizer import NemoBertTokenizer
-from nemo.collections.nlp.datasets.token_classification_dataset import BertTokenClassificationDataset
+from nemo.collections.common.tokenizers.bert_tokenizer import NemoBertTokenizer
+from nemo.collections.nlp.data.token_classification_dataset import BertTokenClassificationDataset
 
 __all__ = ['NERModel']
 
-ACT2FN = {"gelu": nn.functional.gelu, "relu": nn.functional.relu}
+from nemo.collections.nlp.modules.common import TokenClassifier
+
+
 
 
 class NERModel(pl.LightningModule):
@@ -144,7 +144,7 @@ class NERModel(pl.LightningModule):
         if optimizer == 'adam':
             self.__optimizer = torch.optim.Adam(self.parameters(), lr=optim_params['lr'])
         else:
-            raise ValueError(f'TODO {optimizer}')
+            raise NotImplementedError()
 
     def __setup_dataloader_ner(
         self,
@@ -189,33 +189,3 @@ class NERModel(pl.LightningModule):
         return self.__val_dl
 
 
-class TokenClassifier(nn.Module):
-    def __init__(
-        self,
-        hidden_size: object,
-        num_classes: object,
-        activation: object = 'relu',
-        log_softmax: object = True,
-        dropout: object = 0.0,
-        use_transformer_pretrained: object = True,
-    ) -> object:
-        super().__init__()
-        if activation not in ACT2FN:
-            raise ValueError(f'activation "{activation}" not found')
-        self.dense = nn.Linear(hidden_size, hidden_size)
-        self.act = ACT2FN[activation]
-        self.norm = nn.LayerNorm(hidden_size, eps=1e-12)
-        self.mlp = MultiLayerPerceptron(
-            hidden_size, num_classes, num_layers=1, activation=activation, log_softmax=log_softmax
-        )
-        self.dropout = nn.Dropout(dropout)
-        if use_transformer_pretrained:
-            self.apply(lambda module: transformer_weights_init(module, xavier=False))
-
-    def forward(self, hidden_states):
-        hidden_states = self.dropout(hidden_states)
-        hidden_states = self.dense(hidden_states)
-        hidden_states = self.act(hidden_states)
-        transform = self.norm(hidden_states)
-        logits = self.mlp(transform)
-        return logits
