+++ nemo/collections/nlp/modules/__init__.py
@@ -1,0 +1,1 @@
+from nemo.collections.nlp.modules.common import *
