+++ nemo/collections/nlp/modules/common/__init__.py
@@ -1,0 +1,1 @@
+from nemo.collections.nlp.modules.common.token_classifier import TokenClassifier
